+++ src/containers/Home/HomeComparisonCards/HomeComparisonCards.scss
@@ -1,0 +1,57 @@
+.HomeComparisonCards {
+  display: flex;
+  justify-content: center;
+  margin-top: 60px;
+
+  &__arrow-container {
+    align-items: center;
+    display: flex;
+    justify-content: center;
+    width: 60px;
+  }
+
+  &__card {
+    background: white;
+    border-radius: 12px;
+    box-shadow: 0 20px 40px rgba(11, 38, 63, 0.15);
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    min-height: 390px;
+    padding: 30px 32px 60px;
+    width: 360px;
+
+    img.crypto-logo {
+      height: 30px;
+    }
+  }
+
+  &__stat {
+    align-items: center;
+    display: flex;
+    flex-direction: column;
+    margin-top: 60px;
+
+    .attribute {
+      font-size: 20px;
+      font-weight: 500;
+      margin-top: 12px;
+    }
+
+    .value {
+      font-size: 36px;
+      font-weight: 300;
+    }
+  }
+
+  @media (max-width: 768px) {
+    align-items: center;
+    flex-direction: column;
+
+    &__arrow-container {
+      height: 60px;
+      transform: rotate(90deg);
+      width: auto;
+    }
+  }
+}
