+++ src/containers/Home/HomeInstantTransactions/HomeInstantTransactions.scss
@@ -19,57 +19,17 @@
     z-index: 0;
   }
 
-  &__comparison-card-container {
-    display: flex;
-    justify-content: center;
-    margin-top: 60px;
-
-    .arrow-container {
-      align-items: center;
-      display: flex;
-      justify-content: center;
-      width: 60px;
-    }
-
-    .comparison-card {
-      background: white;
-      border-radius: 12px;
-      box-shadow: 0 20px 40px rgba(11, 38, 63, 0.15);
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      min-height: 390px;
-      padding: 30px 32px 60px;
-      width: 360px;
-
-      img.crypto-logo {
-        height: 30px;
-      }
-
-      .comparison-card-stat {
-        align-items: center;
-        display: flex;
-        flex-direction: column;
-        margin-top: 60px;
-
-        .attribute {
-          font-size: 20px;
-          font-weight: 500;
-          margin-top: 12px;
-        }
-
-        .value {
-          font-size: 36px;
-          font-weight: 300;
-        }
-      }
-    }
+  &__blue-block {
+    background: #131f41;
+    display: none;
+    height: 0;
+    width: 100%;
   }
 
   &__content-container {
-    margin-bottom: -18rem;
+    margin-bottom: -150px;
     max-width: 1148px; // (480px * 2) + 60px gutter + 128px padding
-    padding: 128px 64px;
+    padding: 128px 64px 0;
     width: 100%;
     z-index: 1;
 
@@ -108,18 +68,14 @@
     }
   }
 
-  @media (max-width: 768px) {
-    &__comparison-card-container {
-      align-items: center;
-      flex-direction: column;
-
-      .arrow-container {
-        height: 60px;
-        transform: rotate(90deg);
-        width: auto;
-      }
+  @media (max-width: 1440px) {
+    &__blue-block {
+      display: block;
+      height: 180px;
     }
+  }
 
+  @media (max-width: 768px) {
     &__content-container {
       max-width: 480px;
       padding: 32px 32px 128px;
